TraceSegment moved into dd-trace-api
