+++ dd-java-agent/agent-iast/src/main/java/com/datadog/iast/IastTag.java
@@ -1,6 +1,6 @@
 package com.datadog.iast;
 
-import datadog.trace.api.TraceSegment;
+import datadog.trace.api.internal.TraceSegment;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
 import javax.annotation.Nullable;
