+++ dd-java-agent/agent-iast/src/main/java/com/datadog/iast/Reporter.java
@@ -6,9 +6,9 @@ import com.datadog.iast.model.Vulnerability;
 import com.datadog.iast.model.VulnerabilityBatch;
 import datadog.trace.api.Config;
 import datadog.trace.api.DDTags;
-import datadog.trace.api.TraceSegment;
 import datadog.trace.api.gateway.RequestContext;
 import datadog.trace.api.gateway.RequestContextSlot;
+import datadog.trace.api.internal.TraceSegment;
 import datadog.trace.bootstrap.instrumentation.api.*;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
