+++ dd-java-agent/agent-iast/src/test/groovy/com/datadog/iast/IastTagTest.groovy
@@ -1,6 +1,6 @@
 package com.datadog.iast
 
-import datadog.trace.api.TraceSegment
+import datadog.trace.api.internal.TraceSegment
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan
 import datadog.trace.test.util.DDSpecification
